added token validation in http mappings, better logging

+++ src/main/java/de/streaming/service/Controller/UserSerienController.java
@@ -2,17 +2,13 @@ package de.streaming.service.Controller;
 
 import de.streaming.service.Entity.Serie;
 import de.streaming.service.Entity.User;
-import de.streaming.service.Entity.UserSerie;
 import de.streaming.service.Model.UserSerieIds;
 import de.streaming.service.Service.DbService;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.Key;
 import java.util.List;
 
 @RestController
@@ -27,29 +23,44 @@ public class UserSerienController {
 
     @GetMapping(value = "/serie/refresh/all", produces = MediaType.APPLICATION_JSON_VALUE)
     public List<Serie> refreshAS(@RequestHeader("Authorization") String token) {
-        return dbService.refreshSerien();
+        if (dbService.validateToken(token)) {
+            return dbService.refreshSerien();
+        }
+        return null;
     }
 
     @GetMapping(value = "serie/user/refresh/{userId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public List<Serie> refreshUS(@PathVariable Integer userId, @RequestHeader("Authorization") String token) {
-        return dbService.refreshUserSerien(userId);
+        if (dbService.validateToken(token)) {
+            return dbService.refreshUserSerien(userId);
+        }
+        return null;
     }
 
     @GetMapping(value = "user/serie/refresh/same/{serieId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public List<User> refreshSV(@PathVariable Integer serieId, @RequestHeader("Authorization") String token) {
-        return dbService.refreshSameViewers(serieId);
+        if (dbService.validateToken(token)) {
+            return dbService.refreshSameViewers(serieId);
+        }
+        return null;
     }
 
     @PostMapping(value = "serie/user/save", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public HttpStatus saveUS(@RequestBody UserSerieIds userSerieIds, @RequestHeader("Authorization") String token) {
-        dbService.saveUserSerie(userSerieIds.getUserId(), userSerieIds.getSerieId());
-        return HttpStatus.ACCEPTED;
+        if (dbService.validateToken(token)) {
+            dbService.saveUserSerie(userSerieIds.getUserId(), userSerieIds.getSerieId());
+            return HttpStatus.ACCEPTED;
+        }
+        return HttpStatus.UNAUTHORIZED;
     }
 
     @PostMapping(value = "serie/user/delete", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public HttpStatus deleteUS(@RequestBody UserSerieIds userSerieIds, @RequestHeader("Authorization") String token) {
-        dbService.deleteUserSerie(userSerieIds.getUserId(), userSerieIds.getSerieId());
-        return  HttpStatus.ACCEPTED;
+        if (dbService.validateToken(token)) {
+            dbService.deleteUserSerie(userSerieIds.getUserId(), userSerieIds.getSerieId());
+            return HttpStatus.ACCEPTED;
+        }
+        return HttpStatus.UNAUTHORIZED;
     }
